+++ src/control/BarrettFingerSpreadCommandExecutor.cpp
@@ -34,7 +34,7 @@ BarrettFingerSpreadCommandExecutor::BarrettFingerSpreadCommandExecutor(
     }
 
     const auto numDofs = mFingers[i]->getNumDofs();
-    if (_spread >= numDofs)
+    if ((unsigned int)_spread >= numDofs)
       throw std::invalid_argument("Finger does not have spread dof.");
 
     mSpreadDofs.push_back(mFingers[i]->getDof(_spread));
